Creates page-controller and moves render logic there
